+++ src/main/java/de/samply/converter/EmptySession.java
@@ -1,0 +1,9 @@
+package de.samply.converter;
+
+public class EmptySession {
+
+  public static EmptySession instance(){
+    return new EmptySession();
+  }
+
+}