+++ src/main/java/de/samply/excel/Session.java
@@ -1,0 +1,11 @@
+package de.samply.excel;
+
+public class Session {
+
+  private String writeDirectory;
+
+  public Session(String writeDirectory) {
+    this.writeDirectory = writeDirectory;
+  }
+
+}