+++ src/main/java/de/samply/fhir/BundleToContainersConverterSession.java
@@ -1,0 +1,5 @@
+package de.samply.fhir;
+
+public class BundleToContainersConverterSession {
+
+}