+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - MDR URN in attribute template
 - Child fhir path
 - Spring Boot 2.7.7
+- Converter Session
+- Apache POI
 
 ### Changed
 - Rename Conversion Template as Converter Template